requirements.txt
# Core ML frameworks
torch==2.8.0
torchvision==0.23.0
pytorch-lightning
lightning
numpy==1.24.4

# Language Models & Diffusion
transformers==4.51.3
diffusers==0.34.0
accelerate==1.1.1
tokenizers==0.21.4
deepspeed==0.15.1  # Install with: DS_BUILD_UTILS=1 DS_BUILD_FUSED_ADAM=1 DS_BUILD_CPU_ADAM=1 pip install deepspeed
#flash-attn==2.3.5  # Optional, for efficient attention. Install with: pip install flash-attn --no-build-isolation

# Configuration & Logging
hydra-core==1.3.2
omegaconf==2.3.0
pyyaml==6.0.2

# Experiment Tracking
wandb==0.19.1
wandb-osh==1.2.2
tensorboard>=2.15,<2.16

# Media Processing
opencv-python==4.9.0.80
decord==0.6.0
imageio==2.34.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
av==14.0.1  # Note: Changed from pyav==14.0.1
scikit-video==1.1.11

# Image Processing
pillow==11.1.0
einops==0.8.0

# Data Analysis
pandas==2.1.4
scikit-learn==1.5.0
matplotlib==3.8.2
plotly==5.18.0

# Dataset Dependencies
tensorflow==2.15.0  # For tensorflow-datasets
tensorflow-datasets==4.9.7
gcsfs==2024.12.0  # For Google Cloud Storage datasets
ego4d==1.7.3  # For Ego4D dataset
ijson==3.3.0  # For streaming JSON parsing

# Text Processing and Tokenization
sentencepiece==0.2.0
ftfy==6.3.1  # For text cleaning

# Hugging Face Hub
huggingface-hub==0.33.4

# Utilities
tqdm==4.66.2
colorama==0.4.6
click==8.1.8
easydict==1.13
msgpack==1.1.2  # For message serialization
pyzmq==27.1.0  # For ZeroMQ (used in serving)
pydantic

# Optional: Gradio for Demos
gradio==5.11.0